[WIP] Journey Roadmap Section Public Transport : Line Diagram refactoring
